refactor(profile): rename package prof → profile for standard naming
- prof/ directory renamed to profile/
- package prof → package profile
- import path updated to github.com/siyul-park/minivm/profile
- removed profpkg alias; profile.Profile reads naturally
